Move duplicated test classes to file
