+++ test/call_test.py
@@ -1,36 +1,14 @@
 import unittest
 from collections import namedtuple
 
-from parse_this import create_parser, parse_class
+from parse_this import create_parser
 from parse_this.call import (
     _call,
     _call_method_from_namespace,
     _get_parser_call_method,
 )
 from parse_this.exception import ParseThisException
-
-
-@parse_class()
-class Parseable(object):
-    @create_parser()
-    def __init__(self, a: int):
-        self._a = a
-
-    @create_parser()
-    def _private_method(self, b: int):
-        return self._a * b
-
-    def not_parseable(self, c: int):
-        return self._a * c
-
-    @create_parser()
-    def parseable(self, d: int):
-        return self._a * d
-
-    @classmethod
-    @create_parser()
-    def cls_method(cls, e: int):
-        return e * e
+from test.helpers import Parseable
 
 
 def parse_me_no_docstring(one: int, two: int, three: int):
